BitData_V1/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.2
django-allauth==65.1.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
